Extract weavr-tui tests to separate module
Move inline tests from lib.rs to tests.rs submodule to reduce
file size from 602 to 403 lines (33% reduction). No API changes.
